+++ e2e/e2etests/e2etests.go
@@ -98,6 +98,9 @@ const (
 	TestSuiWithdrawName                  = "sui_withdraw"
 	TestSuiTokenWithdrawName             = "sui_token_withdraw" // #nosec G101: Potential hardcoded credentials (gosec), not a credential
 	TestSuiWithdrawAndCallName           = "sui_withdraw_and_call"
+	TestSuiWithdrawRevertWithCallName    = "sui_withdraw_revert_with_call" // #nosec G101: Potential hardcoded credentials (gosec), not a credential
+	TestSuiDepositRestrictedName         = "sui_deposit_restricted"
+	TestSuiWithdrawRestrictedName        = "sui_withdraw_restricted"
 
 	/*
 	 Bitcoin tests
@@ -810,6 +813,15 @@ var AllE2ETests = []runner.E2ETest{
 		runner.WithMinimumVersion("v30.0.0"),
 	),
 	runner.NewE2ETest(
+		TestSuiWithdrawRevertWithCallName,
+		"withdraw SUI from ZEVM that reverts with a onRevert call",
+		[]runner.ArgDefinition{
+			{Description: "amount in mist", DefaultValue: "1000000"},
+		},
+		TestSuiWithdrawRevertWithCall,
+		runner.WithMinimumVersion("v30.0.0"),
+	),
+	runner.NewE2ETest(
 		TestSuiTokenWithdrawName,
 		"withdraw fungible token from ZEVM",
 		[]runner.ArgDefinition{
@@ -817,6 +829,22 @@ var AllE2ETests = []runner.E2ETest{
 		},
 		TestSuiTokenWithdraw,
 		runner.WithMinimumVersion("v29.0.0"),
+	),
+	runner.NewE2ETest(
+		TestSuiDepositRestrictedName,
+		"deposit SUI into ZEVM restricted address",
+		[]runner.ArgDefinition{
+			{Description: "amount in mist", DefaultValue: "1000000"},
+		},
+		TestSuiDepositRestrictedAddress,
+	),
+	runner.NewE2ETest(
+		TestSuiWithdrawRestrictedName,
+		"withdraw SUI from ZEVM to restricted address",
+		[]runner.ArgDefinition{
+			{Description: "amount in mist", DefaultValue: "1000000"},
+		},
+		TestSuiWithdrawRestrictedAddress,
 	),
 	/*
 	 Bitcoin tests