+++ zetaclient/chains/sui/signer/signer.go
@@ -3,6 +3,7 @@ package signer
 import (
 	"context"
 	"crypto/sha256"
+	"fmt"
 
 	"github.com/block-vision/sui-go-sdk/models"
 	"github.com/pkg/errors"
@@ -12,7 +13,7 @@ import (
 	cctypes "github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
-	"github.com/zeta-chain/node/zetaclient/logs"
+	"github.com/zeta-chain/node/zetaclient/compliance"
 )
 
 // Signer Sui outbound transaction signer.
@@ -27,12 +28,17 @@ type Signer struct {
 
 // RPC represents Sui rpc.
 type RPC interface {
+	SuiXGetLatestSuiSystemState(ctx context.Context) (models.SuiSystemStateSummary, error)
 	GetOwnedObjectID(ctx context.Context, ownerAddress, structType string) (string, error)
 
 	MoveCall(ctx context.Context, req models.MoveCallRequest) (models.TxnMetaData, error)
 	SuiExecuteTransactionBlock(
 		ctx context.Context,
 		req models.SuiExecuteTransactionBlockRequest,
+	) (models.SuiTransactionBlockResponse, error)
+	InspectTransactionBlock(
+		ctx context.Context,
+		req models.SuiDevInspectTransactionBlockRequest,
 	) (models.SuiTransactionBlockResponse, error)
 	SuiGetTransactionBlock(
 		ctx context.Context,
@@ -67,17 +73,26 @@ func (s *Signer) ProcessCCTX(ctx context.Context, cctx *cctypes.CrossChainTx, ze
 	s.MarkOutbound(outboundID, true)
 	defer func() { s.MarkOutbound(outboundID, false) }()
 
-	tx, err := s.buildWithdrawal(ctx, cctx)
+	withdrawTxBuilder, err := s.createWithdrawTxBuilder(cctx, zetaHeight)
 	if err != nil {
-		return errors.Wrap(err, "unable to build withdrawal tx")
+		return errors.Wrap(err, "unable to create withdrawal tx builder")
 	}
 
-	sig, err := s.signTx(ctx, tx, zetaHeight, nonce)
+	// always need a cancel tx as fallback
+	cancelTxBuilder, err := s.createCancelTxBuilder(ctx, cctx, zetaHeight)
 	if err != nil {
-		return errors.Wrap(err, "unable to sign tx")
+		return errors.Wrap(err, "unable to create cancel tx builder")
 	}
 
-	txDigest, err := s.broadcast(ctx, tx, sig)
+	var txDigest string
+
+	// broadcast tx according to compliance check result
+	if s.passesCompliance(cctx) {
+		txDigest, err = s.broadcastWithdrawalWithFallback(ctx, withdrawTxBuilder, cancelTxBuilder)
+	} else {
+		txDigest, err = s.broadcastCancelTx(ctx, cancelTxBuilder)
+	}
+
 	if err != nil {
 		// todo we might need additional error handling
 		// for the case when the tx is already broadcasted by another zetaclient
@@ -85,42 +100,112 @@ func (s *Signer) ProcessCCTX(ctx context.Context, cctx *cctypes.CrossChainTx, ze
 		return errors.Wrap(err, "unable to broadcast tx")
 	}
 
-	s.Logger().Std.Info().Str(logs.FieldTx, txDigest).Msg("Broadcasted transaction")
-
-	logger := s.Logger().Std.With().
-		Str(logs.FieldMethod, "reportToOutboundTracker").
-		Int64(logs.FieldChain, s.Chain().ChainId).
-		Uint64(logs.FieldNonce, nonce).
-		Str(logs.FieldTx, txDigest).
-		Logger()
-
-	ctx = logger.WithContext(ctx)
-
-	bg.Work(ctx,
-		func(ctx context.Context) error { return s.reportOutboundTracker(ctx, nonce, txDigest) },
-		bg.WithLogger(logger),
-		bg.WithName("report_outbound_tracker"),
-	)
+	// report outbound tracker
+	task := func(ctx context.Context) error { return s.reportOutboundTracker(ctx, nonce, txDigest) }
+	bg.Work(ctx, task, bg.WithName("report_outbound_tracker"))
 
 	return nil
 }
 
-func (s *Signer) signTx(ctx context.Context, tx models.TxnMetaData, zetaHeight, nonce uint64) ([65]byte, error) {
+// signTx signs a tx with TSS and returns a base64 encoded signature.
+func (s *Signer) signTx(ctx context.Context, tx models.TxnMetaData, zetaHeight, nonce uint64) (string, error) {
 	digest, err := sui.Digest(tx)
 	if err != nil {
-		return [65]byte{}, errors.Wrap(err, "unable to get digest")
+		return "", errors.Wrap(err, "unable to get digest")
 	}
 
-	// Another hashing is required for ECDSA.
-	// https://docs.sui.io/concepts/cryptography/transaction-auth/signatures#signature-requirements
-	digestWrapped := sha256.Sum256(digest[:])
+	// send TSS signature request.
+	sig65B, err := s.TSS().Sign(ctx, wrapDigest(digest), zetaHeight, nonce, s.Chain().ChainId)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to sign tx")
+	}
 
-	// Send TSS signature request.
-	return s.TSS().Sign(
-		ctx,
-		digestWrapped[:],
-		zetaHeight,
-		nonce,
+	// serialize signature
+	sigBase64, err := sui.SerializeSignatureECDSA(sig65B, s.TSS().PubKey().AsECDSA())
+	if err != nil {
+		return "", errors.Wrap(err, "unable to serialize tx signature")
+	}
+
+	return sigBase64, nil
+}
+
+// SignTxWithCancel signs original tx and cancel tx in one go to save TSS keysign time.
+//
+// Note: this function is not used due to tx simulation issue in Sui SDK,
+// but we can sign both tx and cancel tx in one go once Sui SDK is updated.
+func (s *Signer) SignTxWithCancel(
+	ctx context.Context,
+	tx models.TxnMetaData,
+	txCancel models.TxnMetaData,
+	zetaHeight, nonce uint64,
+) (sig string, sigCancel string, err error) {
+	digests := make([][]byte, 2)
+
+	// tx digest
+	digest, err := sui.Digest(tx)
+	if err != nil {
+		return "", "", errors.Wrap(err, "unable to get tx digest")
+	}
+	digests[0] = wrapDigest(digest)
+
+	// cancel tx digest
+	digestCancel, err := sui.Digest(txCancel)
+	if err != nil {
+		return "", "", errors.Wrap(err, "unable to get cancel tx digest")
+	}
+	digests[1] = wrapDigest(digestCancel)
+
+	// sign both digests with TSS
+	sig65Bs, err := s.TSS().SignBatch(ctx, digests, zetaHeight, nonce, s.Chain().ChainId)
+	if err != nil {
+		return "", "", errors.Wrapf(err, "unable to sign %d tx(s) with TSS", len(digests))
+	}
+
+	// should never mismatch
+	if len(sig65Bs) != len(digests) {
+		return "", "", fmt.Errorf("expected %d signatures, got %d", len(digests), len(sig65Bs))
+	}
+
+	// serialize signatures
+	sig, err = sui.SerializeSignatureECDSA(sig65Bs[0], s.TSS().PubKey().AsECDSA())
+	if err != nil {
+		return "", "", errors.Wrap(err, "unable to serialize tx signature")
+	}
+
+	sigCancel, err = sui.SerializeSignatureECDSA(sig65Bs[1], s.TSS().PubKey().AsECDSA())
+	if err != nil {
+		return "", "", errors.Wrap(err, "unable to serialize tx cancel signature")
+	}
+
+	return sig, sigCancel, nil
+}
+
+func (s *Signer) passesCompliance(cctx *cctypes.CrossChainTx) bool {
+	restricted := compliance.IsCCTXRestricted(cctx)
+	if !restricted {
+		return true
+	}
+
+	params := cctx.GetCurrentOutboundParam()
+
+	compliance.PrintComplianceLog(
+		s.Logger().Std,
+		s.Logger().Compliance,
+		true,
 		s.Chain().ChainId,
+		cctx.Index,
+		cctx.InboundParams.Sender,
+		params.Receiver,
+		params.CoinType.String(),
 	)
+
+	return false
+}
+
+// wrapDigest wraps the digest with sha256.
+func wrapDigest(digest [32]byte) []byte {
+	// another hashing is required for ECDSA.
+	// see: https://docs.sui.io/concepts/cryptography/transaction-auth/signatures#signature-requirements
+	digestWrapped := sha256.Sum256(digest[:])
+	return digestWrapped[:]
 }